refactor(ui): Extract a component for plugin options
The forms for creating and editing plugin templates are almost
identical, so extract the common code to a component.
